Fixed scenario where the Fault<TJob> event was not published when UseMessageRetry is configured on the bus
